+++ sm-core/src/main/java/com/salesmanager/core/business/promotion/dao/PromotionTargetAgeDao.java
@@ -3,6 +3,6 @@ package com.salesmanager.core.business.promotion.dao;
 import com.salesmanager.core.business.generic.dao.SalesManagerEntityDao;
 import com.salesmanager.core.business.promo.model.PromotionTragetAge;
 
-public interface PromotionTargetAgeDao extends SalesManagerEntityDao<Long, PromotionTragetAge> {
+public interface PromotionTargetAgeDao extends SalesManagerEntityDao<Integer, PromotionTragetAge> {
 
 }
