+++ sm-core/src/main/java/com/salesmanager/core/business/promotion/dao/PromotionTragetAgeDaoImpl.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Repository;
 import com.salesmanager.core.business.generic.dao.SalesManagerEntityDaoImpl;
 import com.salesmanager.core.business.promo.model.PromotionTragetAge;
 @Repository("promotionTragetAgeDao")
-public class PromotionTragetAgeDaoImpl extends SalesManagerEntityDaoImpl<Long, PromotionTragetAge>
+public class PromotionTragetAgeDaoImpl extends SalesManagerEntityDaoImpl<Integer, PromotionTragetAge>
 		implements PromotionTargetAgeDao {
