+++ sm-core/src/main/java/com/salesmanager/core/business/promotion/service/ProductAgeRangeSerivce.java
@@ -3,6 +3,6 @@ package com.salesmanager.core.business.promotion.service;
 import com.salesmanager.core.business.generic.service.SalesManagerEntityService;
 import com.salesmanager.core.business.promo.model.PromotionTragetAge;
 
-public interface ProductAgeRangeSerivce extends SalesManagerEntityService<Long, PromotionTragetAge> {
+public interface ProductAgeRangeSerivce extends SalesManagerEntityService<Integer, PromotionTragetAge> {
 
 }
